Lab1/requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
category-encoders